+++ src/sparsematrix/Vector.c
@@ -1,6 +1,6 @@
 #include "Vector.h"
 
-Vector *Vector_new(unsigned int size) {
+Vector *Vector_new(u_int64_t size) {
     Vector *vector = malloc(sizeof(Vector));
     vector->size = size;
     vector->data = malloc(size * sizeof(float));
@@ -24,7 +24,7 @@ int Vector_equals(const Vector *v1, const Vector *v2) {
     int areSameBuffer = v1->data == v2->data;
     if (!areSameSize) return 0;
     if (areSameBuffer) return 1;
-    for (int i = 0; i < v1->size; i++) {
+    for (u_int64_t i = 0; i < v1->size; i++) {
         if (fabsf(v1->data[i] - v2->data[i]) > 0.000001) {
             return 0;
         }
@@ -45,28 +45,11 @@ void Vector_outAsJSON(Vector *vector, FILE *out) {
         fprintf(out, "{}");
     }
     fprintf(out, "%s\n", "{ ");
-    fprintf(out, "%s: %d,\n", "\"size\"",  vector->size);
+    fprintf(out, "%s: %lu,\n", "\"size\"",  vector->size);
     fprintf(out, "%s: [ ", "\"data\"");
-    for (int i=0; i < vector->size - 1; i++) {
+    for (u_int64_t i = 0; i < vector->size - 1; i++) {
         fprintf(out, "%f, ", vector->data[i]);
     }
     fprintf(out, "%f ],\n", vector->data[vector->size - 1]);
     fprintf(out, "%s", "}");
 }
-
-//void Vector::swap(SwapMap& swapMap) {
-//    float *temp = new float[size];
-//    for (int i = 0; i < size; i++) {
-//        temp[i] = data[swapMap.getMapping(i)];
-//    }
-//    memcpy(data, temp, size * sizeof(float));
-//    delete[] temp;
-//}
-//void Vector::swapInverse(SwapMap& swapMap) {
-//    float *temp = new float[size];
-//    for (int i = 0; i < size; i++) {
-//        temp[swapMap.getInverse(i)] = data[i];
-//    }
-//    memcpy(data, temp, size * sizeof(float));
-//    delete[] temp;
-//}
